refactor(Util): Move util, vfile, & puppeteer files to src/util folder
