+++ src/codecs/latex/index.ts
@@ -5,7 +5,7 @@
 import stencila from '@stencila/schema'
 import { Encode } from '../..'
 import * as pandoc from '../pandoc'
-import * as vfile from '../../vfile'
+import * as vfile from '../../util/vfile'
 
 export const mediaTypes = ['application/x-latex']
