+++ src/codecs/pandoc/index.ts
@@ -4,9 +4,9 @@ import childProcess from 'child_process'
 import tempy from 'tempy'
 import { Encode, EncodeOptions } from '../..'
 import { pandocDataDir, pandocPath } from '../../boot'
-import { wrapInBlockNode } from '../../util'
+import { wrapInBlockNode } from '../../util/index'
 import type from '../../util/type'
-import * as vfile from '../../vfile'
+import * as vfile from '../../util/vfile'
 import * as rpng from '../rpng'
 import * as Pandoc from './types'
