+++ src/process.ts
@@ -2,9 +2,9 @@ import stencila from '@stencila/schema'
 import assert from 'assert'
 import path from 'path'
 import { dump, load, read, write } from '.'
-import { validate } from './util'
+import { validate } from './util/index'
 import type from './util/type'
-import { isPath } from './vfile'
+import { isPath } from './util/vfile'
 
 /**
  * Process a node
